Dockerfile: Use /opaal dir owned by nobody so opaal can write
Without this, opaal, running as nobody, tries to write its sqlite db cache file
to / which is owned by root. With this, an /opaal directory is created, owned
by nobody, and the opaal binary is copied here. That way, it has permission to
write its cache file here. Any config can be copied here as well.